refactored the collection_use_cases to include a collections namespace and a less verbose file name that uses the context of the namespace
