+++ controllers/collections_controller.py
@@ -1,5 +1,5 @@
 from cmd import Cmd
-from application.use_cases.collections_use_cases import CollectionsUseCases
+from application.use_cases.collections.use_cases import CollectionsUseCases
 from common.guid_utilities import is_uuid4, to_uuid4
 
 class CollectionsController(Cmd):
